+++ src/index.js
@@ -6,6 +6,7 @@ export * from './components/FormFields';
 export * from './components/IconButtons';
 export * from './components/LoadingIndicator';
 export * from './components/CheckIndicator';
+export * from './components/SearchBox';
 
 export * from './utils/Hooks';
 export * from './utils/TimeAndDate';