+++ src/utils/Hooks.js
@@ -2,12 +2,12 @@ import React, { useRef, useEffect, useState, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
 export const usePrevious = (value) => {
-    const ref = useRef();
-    useEffect(() => {
-      ref.current = value;
-    });
-    return ref.current;
-  }
+  const ref = useRef();
+  useEffect(() => {
+    ref.current = value;
+  });
+  return ref.current;
+};
   
 // https://stackoverflow.com/questions/54666401/how-to-use-throttle-or-debounce-with-react-hook
 export const useDebouncedEffect = (effect, deps, delay) => {